+++ allennlp/training/callbacks/__init__.py
@@ -1,3 +1,14 @@
 from allennlp.training.callbacks.callback import Callback, handle_event
 from allennlp.training.callbacks.callback_handler import CallbackHandler
 from allennlp.training.callbacks.events import Events
+
+from allennlp.training.callbacks.log_to_tensorboard import LogToTensorboard
+from allennlp.training.callbacks.learning_rate_scheduler import LrsCallback
+from allennlp.training.callbacks.momentum_scheduler import MomentumSchedulerCallback
+from allennlp.training.callbacks.checkpoint import CheckpointCallback
+from allennlp.training.callbacks.moving_average import MovingAverageCallback
+from allennlp.training.callbacks.validate import Validate
+from allennlp.training.callbacks.track_metrics import TrackMetrics
+from allennlp.training.callbacks.train_supervised import TrainSupervised
+from allennlp.training.callbacks.generate_training_batches import GenerateTrainingBatches
+from allennlp.training.callbacks.post_to_url import PostToUrl
