+++ allennlp/training/callbacks/events.py
@@ -1,28 +1,22 @@
 class Events:
     TRAINING_START = "TRAINING_START"
+
     EPOCH_START = "EPOCH_START"
 
     BATCH_START = "BATCH_START"
 
-    BEFORE_FORWARD = "BEFORE_FORWARD"
     FORWARD = "FORWARD"
-    AFTER_FORWARD = "AFTER_FORWARD"
 
-    BEFORE_BACKWARD = "BEFORE_BACKWARD"
     BACKWARD = "BACKWARD"
-    AFTER_BACKWARD = "AFTER_BACKWARD"
 
     BATCH_END = "BATCH_END"
 
-    BEFORE_VALIDATE = "BEFORE_VALIDATE"
     VALIDATE = "VALIDATE"
-    AFTER_VALIDATE = "AFTER_VALIDATE"
 
     EPOCH_END = "EPOCH_END"
+
     TRAINING_END = "TRAINING_END"
 
-    BEFORE_SAVE_CHECKPOINT = "BEFORE_SAVE_CHECKPOINT"
     SAVE_CHECKPOINT = "SAVE_CHECKPOINT"
-    AFTER_SAVE_CHECKPOINT = "AFTER_SAVE_CHECKPOINT"
 
     RESTORE_CHECKPOINT = "RESTORE_CHECKPOINT"
