+++ doc/api/allennlp.training.callbacks.rst
@@ -16,7 +16,53 @@ allennlp.training.callbacks
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: allennlp.training.callbacks.callbacks
+
+.. automodule:: allennlp.training.callbacks.checkpoint
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.generate_training_batches
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.learning_rate_scheduler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.log_to_tensorboard
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.momentum_scheduler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.moving_average
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.post_to_url
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.track_metrics
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.train_supervised
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.callbacks.validate
    :members:
    :undoc-members:
    :show-inheritance:
